% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmt.R
\name{fmt1Decimale}
\alias{fmt1Decimale}
\title{fmt1Decimale}
\usage{
fmt1Decimale(x)
}
\arguments{
\item{x}{un nombre}
}
\value{
Le nombre formate
}
\description{
Formate un nombre
}
\details{
La fonction formate avec
une decimale, 
la virgule pour distinguer unites et decimales, 
l'espace comme separateur de milliers.
}
\section{Emploie }{

Les fonctions \code{\link{fmt}} et \url{https://www.rdocumentation.org/packages/base/versions/3.6.2/topics/format}
}

\section{Warning }{

N'utilise pas de notation scientifique
}

\examples{
fmt1Decimale(100)
fmt1Decimale(1234.56789)
}
\references{
\url{https://www.rdocumentation.org/packages/base/versions/3.6.2/topics/format} la fonction format
}
\seealso{
\url{https://www.rdocumentation.org/packages/base/versions/3.6.2/topics/print} pour afficher dans la console

Other fonctions pour formater: 
\code{\link{fmt}()}
}
\author{
M. Martin, \email{martin@insee.fr}
}
\concept{fonctions pour formater}
